fixed Api. Removed foreign key. fixed metadatarepository address

+++ core/data/src/main/java/com/core/data/repository/Web3jNetworkBalanceRepository.kt
@@ -59,7 +59,7 @@ class Web3jNetworkBalanceRepository @Inject constructor(
                             tokenBalanceDao.upsertTokenBalances(
                                 listOf(
                                     tokenBalance.copy(
-                                        tokenBalance = newNetworkBalance.toDouble()
+                                        tokenBalance = newNetworkBalance
                                     )
                                 )
                             )
