+++ core/model/src/main/java/com/core/model/TokenBalance.kt
@@ -1,7 +1,9 @@
 package com.core.model
 
+import java.math.BigDecimal
+
 data class TokenBalance(
     val contractAddress: String,
     val chainId: Int,
-    val tokenBalance: Double,
+    val tokenBalance: BigDecimal,
 )
